+++ rust/bridge/shared/types/src/keytrans.rs
@@ -22,17 +22,3 @@ impl libsignal_net_chat::ws::WsConnection for crate::net::chat::UnauthenticatedC
         BridgeChatConnection::send(self, request, Duration::MAX)
     }
 }
-
-/// A wrapper around libsignal-net-chat's error scheme to stay compatible with existing KT APIs as
-/// exposed at the app layer.
-#[derive(Debug, derive_more::From, derive_more::Into, derive_more::Deref, thiserror::Error)]
-#[error("{0}")]
-pub struct BridgeError(
-    libsignal_net_chat::api::RequestError<libsignal_net_chat::api::keytrans::Error>,
-);
-
-impl From<libsignal_net_chat::api::keytrans::Error> for BridgeError {
-    fn from(inner: libsignal_net_chat::api::keytrans::Error) -> Self {
-        Self(libsignal_net_chat::api::RequestError::Other(inner))
-    }
-}