+++ RELEASE_NOTES.md
@@ -1,2 +1,3 @@
 v0.82.1
 
+- keytrans: Unify errors with other typed APIs